# docs/swagger.yaml
basePath: /
definitions:
  api.CreatureModel:
    properties:
      attack:
        type: integer
      defence:
        type: integer
      description:
        type: string
      hp:
        type: integer
      id:
        type: string
      name:
        type: string
      teach_spell:
        items:
          $ref: '#/definitions/api.SpellModel'
        type: array
    type: object
  api.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
  api.SpellModel:
    properties:
      description:
        type: string
      element:
        type: string
      id:
        type: string
      mana_cost:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API De Magias e Criaturas
  title: Grimorio API
  version: "1.0"
paths:
  /api/creature:
    get:
      description: Lista todas as crituras com filtros opcionais e paginação
      parameters:
      - description: nome da criatura
        in: query
        name: name
        type: string
      - description: Ataque
        in: query
        name: attack
        type: integer
      - description: Defesa
        in: query
        name: defence
        type: integer
      - description: Campo de ordenação (name, attack)
        in: query
        name: sort
        type: string
      - description: asc ou desc
        in: query
        name: order
        type: string
      - description: Pagina
        in: query
        name: page
        type: integer
      - description: limite
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      summary: Lista todas as criaturas
      tags:
      - creatures
    post:
      consumes:
      - application/json
      description: Cria uma nova criatura
      parameters:
      - description: Dados da criatura
        in: body
        name: creature
        required: true
        schema:
          $ref: '#/definitions/api.CreatureModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      summary: Cria uma criatura
      tags:
      - creatures
  /api/creature/{id}:
    delete:
      description: Deleta uma criatura do grimorio
      parameters:
      - description: ID da criatura
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Deleta criatura
      tags:
      - creatures
    get:
      description: Retorna uma criaturas especifica
      parameters:
      - description: ID da criatura
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Busca criatura por ID
      tags:
      - creatures
    put:
      consumes:
      - application/json
      description: Atualiza dados da criatura
      parameters:
      - description: ID da criatura
        in: path
        name: id
        type: string
      - description: Dados Atualizados
        in: body
        name: creature
        required: true
        schema:
          $ref: '#/definitions/api.CreatureModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Atualiza criatura
      tags:
      - creatures
  /api/creature/{id}/teach/{spellId}:
    post:
      description: Associa uma magina a uma Criatura
      parameters:
      - description: ID da criatura
        in: path
        name: id
        type: string
      - description: ID da magia
        in: path
        name: spellId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      summary: Criatura ensina uma nova magia
      tags:
      - creatures
  /api/spell:
    get:
      parameters:
      - description: Nome da magia
        in: query
        name: name
        type: string
      - description: elemento da magia
        in: query
        name: element
        type: string
      - description: name ou mana_cost
        in: query
        name: sort
        type: string
      - description: asc ou desc
        in: query
        name: order
        type: string
      - description: Página
        in: query
        name: page
        type: integer
      - description: Limite
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      summary: Lista magias com filtros
      tags:
      - spells
    post:
      consumes:
      - application/json
      description: Cria uma magia
      parameters:
      - description: Dados da magia
        in: body
        name: spell
        required: true
        schema:
          $ref: '#/definitions/api.SpellModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      summary: Cria uma nova magia
      tags:
      - spells
  /api/spell/{id}:
    delete:
      description: Deleta magia do grimorio
      parameters:
      - description: ID da magia
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Deleta magia
      tags:
      - spells
    get:
      parameters:
      - description: ID da magia
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Busca magias por ID
      tags:
      - spells
    put:
      consumes:
      - application/json
      description: Atualiza dados da magia
      parameters:
      - description: ID da magia
        in: path
        name: id
        required: true
        type: string
      - description: Dados Atualizados
        in: body
        name: spell
        required: true
        schema:
          $ref: '#/definitions/api.SpellModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      summary: Atualiza magias
      tags:
      - spells
  /api/spell/element/{element}:
    get:
      description: Lista magias por elemento
      parameters:
      - description: Elemento da magia
        in: path
        name: element
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      summary: Lista magias por elemento
      tags:
      - spells
swagger: "2.0"